+++ jobs/knative-job-server/src/run.ts
@@ -3,11 +3,12 @@
 import server from './index';
 import poolManager from '@constructive-io/job-pg';
 import { getJobsCallbackPort } from '@constructive-io/job-utils';
+import { createLogger } from '@pgpmjs/logger';
 
+const logger = createLogger('knative-job-server');
 const pgPool = poolManager.getPool();
 const port = getJobsCallbackPort();
 
 server(pgPool).listen(port, () => {
-  // eslint-disable-next-line no-console
-  console.log(`listening ON ${port}`);
+  logger.info(`listening ON ${port}`);
 });