+++ functions/simple-email/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.3.0](https://github.com/constructive-io/constructive/compare/@constructive-io/simple-email-fn@0.2.22...@constructive-io/simple-email-fn@0.3.0) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/simple-email-fn
+
+## [0.2.22](https://github.com/constructive-io/constructive/compare/@constructive-io/simple-email-fn@0.2.21...@constructive-io/simple-email-fn@0.2.22) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/simple-email-fn
+
 ## [0.2.21](https://github.com/constructive-io/constructive/compare/@constructive-io/simple-email-fn@0.2.20...@constructive-io/simple-email-fn@0.2.21) (2026-01-09)
 
 **Note:** Version bump only for package @constructive-io/simple-email-fn