+++ graphile/graphile-many-to-many/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.3.0](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.17...graphile-many-to-many@1.3.0) (2026-01-18)
+
+**Note:** Version bump only for package graphile-many-to-many
+
+## [1.2.17](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.16...graphile-many-to-many@1.2.17) (2026-01-18)
+
+**Note:** Version bump only for package graphile-many-to-many
+
 ## [1.2.16](https://github.com/constructive-io/constructive/compare/graphile-many-to-many@1.2.15...graphile-many-to-many@1.2.16) (2026-01-14)
 
 **Note:** Version bump only for package graphile-many-to-many