+++ graphile/graphile-postgis/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/constructive-io/constructive/compare/graphile-postgis@0.4.17...graphile-postgis@0.5.0) (2026-01-18)
+
+**Note:** Version bump only for package graphile-postgis
+
+## [0.4.17](https://github.com/constructive-io/constructive/compare/graphile-postgis@0.4.16...graphile-postgis@0.4.17) (2026-01-18)
+
+**Note:** Version bump only for package graphile-postgis
+
 ## [0.4.16](https://github.com/constructive-io/constructive/compare/graphile-postgis@0.4.15...graphile-postgis@0.4.16) (2026-01-14)
 
 **Note:** Version bump only for package graphile-postgis