+++ graphile/graphile-simple-inflector/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.5.0](https://github.com/constructive-io/constructive/compare/graphile-simple-inflector@0.4.18...graphile-simple-inflector@0.5.0) (2026-01-18)
+
+**Note:** Version bump only for package graphile-simple-inflector
+
+## [0.4.18](https://github.com/constructive-io/constructive/compare/graphile-simple-inflector@0.4.17...graphile-simple-inflector@0.4.18) (2026-01-18)
+
+**Note:** Version bump only for package graphile-simple-inflector
+
 ## [0.4.17](https://github.com/constructive-io/constructive/compare/graphile-simple-inflector@0.4.16...graphile-simple-inflector@0.4.17) (2026-01-14)
 
 **Note:** Version bump only for package graphile-simple-inflector