+++ graphile/graphile-test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.14.0](https://github.com/constructive-io/constructive/compare/graphile-test@2.13.17...graphile-test@2.14.0) (2026-01-18)
+
+**Note:** Version bump only for package graphile-test
+
+## [2.13.17](https://github.com/constructive-io/constructive/compare/graphile-test@2.13.16...graphile-test@2.13.17) (2026-01-18)
+
+**Note:** Version bump only for package graphile-test
+
 ## [2.13.16](https://github.com/constructive-io/constructive/compare/graphile-test@2.13.15...graphile-test@2.13.16) (2026-01-14)
 
 **Note:** Version bump only for package graphile-test