+++ graphql/env/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-env@2.8.20...@constructive-io/graphql-env@2.9.0) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/graphql-env
+
+## [2.8.20](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-env@2.8.19...@constructive-io/graphql-env@2.8.20) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/graphql-env
+
 ## [2.8.19](https://github.com/constructive-io/constructive/compare/@constructive-io/graphql-env@2.8.18...@constructive-io/graphql-env@2.8.19) (2026-01-11)
 
 **Note:** Version bump only for package @constructive-io/graphql-env