+++ jobs/knative-job-service/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.8.1](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.8.0...@constructive-io/knative-job-service@0.8.1) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+# [0.8.0](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.7.0...@constructive-io/knative-job-service@0.8.0) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+# [0.7.0](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.16...@constructive-io/knative-job-service@0.7.0) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
+## [0.6.16](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.15...@constructive-io/knative-job-service@0.6.16) (2026-01-18)
+
+**Note:** Version bump only for package @constructive-io/knative-job-service
+
 ## [0.6.15](https://github.com/constructive-io/jobs/compare/@constructive-io/knative-job-service@0.6.14...@constructive-io/knative-job-service@0.6.15) (2026-01-14)
 
 **Note:** Version bump only for package @constructive-io/knative-job-service