+++ packages/server-utils/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.9.0](https://github.com/constructive-io/constructive/compare/@pgpmjs/server-utils@2.8.16...@pgpmjs/server-utils@2.9.0) (2026-01-18)
+
+**Note:** Version bump only for package @pgpmjs/server-utils
+
+## [2.8.16](https://github.com/constructive-io/constructive/compare/@pgpmjs/server-utils@2.8.15...@pgpmjs/server-utils@2.8.16) (2026-01-18)
+
+**Note:** Version bump only for package @pgpmjs/server-utils
+
 ## [2.8.15](https://github.com/constructive-io/constructive/compare/@pgpmjs/server-utils@2.8.14...@pgpmjs/server-utils@2.8.15) (2026-01-09)
 
 **Note:** Version bump only for package @pgpmjs/server-utils