+++ postgres/drizzle-orm-test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.6.0](https://github.com/constructive-io/constructive/compare/drizzle-orm-test@0.5.16...drizzle-orm-test@0.6.0) (2026-01-18)
+
+**Note:** Version bump only for package drizzle-orm-test
+
+## [0.5.16](https://github.com/constructive-io/constructive/compare/drizzle-orm-test@0.5.15...drizzle-orm-test@0.5.16) (2026-01-18)
+
+**Note:** Version bump only for package drizzle-orm-test
+
 ## [0.5.15](https://github.com/constructive-io/constructive/compare/drizzle-orm-test@0.5.14...drizzle-orm-test@0.5.15) (2026-01-14)
 
 **Note:** Version bump only for package drizzle-orm-test