+++ postgres/introspectron/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.17.0](https://github.com/constructive-io/constructive/compare/introspectron@2.16.18...introspectron@2.17.0) (2026-01-18)
+
+**Note:** Version bump only for package introspectron
+
+## [2.16.18](https://github.com/constructive-io/constructive/compare/introspectron@2.16.17...introspectron@2.16.18) (2026-01-18)
+
+**Note:** Version bump only for package introspectron
+
 ## [2.16.17](https://github.com/constructive-io/constructive/compare/introspectron@2.16.16...introspectron@2.16.17) (2026-01-14)
 
 **Note:** Version bump only for package introspectron