+++ uploads/uuid-stream/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.4.0](https://github.com/constructive-io/constructive/compare/uuid-stream@2.3.6...uuid-stream@2.4.0) (2026-01-18)
+
+**Note:** Version bump only for package uuid-stream
+
 ## [2.3.6](https://github.com/constructive-io/constructive/compare/uuid-stream@2.3.5...uuid-stream@2.3.6) (2026-01-02)
 
 **Note:** Version bump only for package uuid-stream